+++ Bugsnag/Payload/BugsnagNotifier.m
@@ -21,7 +21,7 @@
 #else
         _name = @"Bugsnag Objective-C";
 #endif
-        _version = @"6.16.8";
+        _version = @"6.17.0";
         _url = @"https://github.com/bugsnag/bugsnag-cocoa";
         _dependencies = @[];
     }